🚐 Move test db files to correct folder
